fixes/xcode adaptation

+++ Deserialize.h
@@ -8,7 +8,7 @@ namespace MeshLoader {
 	using tag_t = uint32_t;
 	using index_t = uint32_t;
 	using poly_t = uint16_t;
-	constexpr uint8_t alignment = sizeof(size_t);
+	constexpr uint8_t alignment = sizeof(uint64_t);
 	struct Header {
 		uint32_t magicNumber, version;
 		uint8_t alignment, tagSize, indexSize/*size for indexing into arrays*/, polySize /*polygon indices*/;
